requirements.txt
biplist
PyYAML
PyDrive2
matplotlib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib